+++ monocliche/src/actions/GoToBoxAction.py
@@ -1,6 +1,6 @@
 from uuid import UUID
 
-from monocliche import Constants
+from monocliche.src import Constants
 
 from monocliche.model.AbstractAction import AbstractAction
