+++ monocliche/tests/actions/GoToJailActionTest.py
@@ -1,6 +1,6 @@
 import unittest
 
-from monocliche import Constants
+from monocliche.src import Constants
 
 from monocliche.model.Board import Board
 from monocliche.model.Box import Box
